[refactor] Moved Splat Modal Views to dedicated modalViews folder; Added additional links in DashHeader for Gallery and Splat Management; Decoupled Delete Form for reusability
